microservices/ServicioFacturacion/requirements.txt
flask
flask-Cors
flask-JWT-Extended
flask-JWT
flask-RESTful
flask-SQLAlchemy
marshmallow_sqlalchemy
python-dotenv
coverage
coverage-lcov 
boto3
python-dotenv
psycopg2-binary==2.9.6
pytest
pytest-cov
pytest-mock
pytest-env
pyjwt
cryptography
requests